reorganize mock logger
