+++ bctl/agent/plugin/kube/actions/portforward/portforward_test.go
@@ -10,9 +10,10 @@ import (
 	"testing"
 	"time"
 
-	"bastionzero.com/bctl/v1/bzerolib/mocks"
+	"bastionzero.com/bctl/v1/bzerolib/logger"
 	"bastionzero.com/bctl/v1/bzerolib/plugin/kube/actions/portforward"
 	smsg "bastionzero.com/bctl/v1/bzerolib/stream/message"
+	"bastionzero.com/bctl/v1/bzerolib/tests"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/tomb.v2"
 	"k8s.io/apimachinery/pkg/util/httpstream"
@@ -67,7 +68,7 @@ func buildStopActionPayload(requestId string) []byte {
 }
 
 // inject our mocked object
-func setDoDial(streamConnection *mocks.MockStreamConnection) {
+func setDoDial(streamConnection *tests.MockStreamConnection) {
 	doDial = func(dialer httpstream.Dialer, protocolName string) (httpstream.Connection, string, error) {
 		return streamConnection, "", nil
 	}
@@ -93,7 +94,7 @@ func TestMain(m *testing.M) {
 
 func TestPortforward(t *testing.T) {
 	assert := assert.New(t)
-	logger := mocks.MockLogger()
+	logger := logger.MockLogger()
 	var tmb tomb.Tomb
 	outputChan := make(chan smsg.StreamMessage, 1)
 
@@ -101,12 +102,12 @@ func TestPortforward(t *testing.T) {
 	portForwardRequestId := "pid"
 	testData := "test data"
 
-	mockStream := mocks.MockStream{MyStreamData: testData}
+	mockStream := tests.MockStream{MyStreamData: testData}
 	mockStream.On("Read", make([]byte, portforward.DataStreamBufferSize)).Return(9, nil)
 	mockStream.On("Write", []byte(testData)).Return(len(testData), nil)
 	mockStream.On("Close").Return(nil)
 
-	mockStreamConnection := new(mocks.MockStreamConnection)
+	mockStreamConnection := new(tests.MockStreamConnection)
 	mockStreamConnection.On("CreateStream", http.Header{
 		"Port":      []string{"5000"},
 		"Requestid": []string{portForwardRequestId},
