+++ bctl/daemon/plugin/kube/actions/stream/stream_test.go
@@ -8,9 +8,10 @@ import (
 	"testing"
 	"time"
 
-	"bastionzero.com/bctl/v1/bzerolib/mocks"
+	"bastionzero.com/bctl/v1/bzerolib/logger"
 	"bastionzero.com/bctl/v1/bzerolib/plugin/kube/actions/stream"
 	smsg "bastionzero.com/bctl/v1/bzerolib/stream/message"
+	"bastionzero.com/bctl/v1/bzerolib/tests"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/tomb.v2"
 )
@@ -27,7 +28,7 @@ func TestMain(m *testing.M) {
 func TestStream(t *testing.T) {
 	assert := assert.New(t)
 	var tmb tomb.Tomb
-	logger := mocks.MockLogger()
+	logger := logger.MockLogger()
 	requestId := "rid"
 	logId := "lid"
 	command := "logs"
@@ -37,9 +38,9 @@ func TestStream(t *testing.T) {
 	receiveData4 := "receive data 4"
 	urlPath := "test-path"
 
-	request := mocks.MockHttpRequest("GET", urlPath, make(map[string][]string), sendData)
+	request := tests.MockHttpRequest("GET", urlPath, make(map[string][]string), sendData)
 
-	writer := mocks.MockResponseWriter{}
+	writer := tests.MockResponseWriter{}
 	writer.On("Write", []byte(receiveData1)).Return(12, nil)
 	writer.On("Write", []byte(receiveData2)).Return(14, nil)
 	writer.On("Header").Return(make(map[string][]string))
