+++ bzerolib/logger/mocks.go
@@ -1,0 +1,8 @@
+package logger
+
+func MockLogger() *Logger {
+	if logger, err := New(DefaultLoggerConfig(Debug.String()), "/dev/null", false); err == nil {
+		return logger
+	}
+	return nil
+}
