+++ bzerolib/tests/mocks.go
@@ -1,30 +1,18 @@
-package mocks
+package tests
 
 import (
 	"bytes"
-	"encoding/base64"
 	"io"
 	"io/ioutil"
 	"net/http"
 	"net/url"
 	"time"
 
-	"bastionzero.com/bctl/v1/bzerolib/logger"
-	smsg "bastionzero.com/bctl/v1/bzerolib/stream/message"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"k8s.io/apimachinery/pkg/util/httpstream"
 )
 
-func MockLogger() *logger.Logger {
-	logger, err := logger.New(logger.DefaultLoggerConfig(logger.Debug.String()), "/dev/null", false)
-	if err == nil {
-		return logger
-	}
-	return nil
-}
-
-// FIXME: docstring
+// MockResponseWriter can be injected into code that writes output (usually daemon-side)
 type MockResponseWriter struct {
 	mock.Mock
 	http.ResponseWriter
@@ -44,7 +32,30 @@ func (m MockResponseWriter) WriteHeader(statusCode int) {
 	m.Called(statusCode)
 }
 
-// FIXME: docstring
+// MockStreamConnection can be injected into code that reads from or writes to an http stream
+// when implemented it should usually return a MockStream from calls to CreateStream
+type MockStreamConnection struct {
+	mock.Mock
+	httpstream.Connection
+}
+
+func (m *MockStreamConnection) CreateStream(headers http.Header) (httpstream.Stream, error) {
+	args := m.Called(headers)
+	return args.Get(0).(httpstream.Stream), args.Error(1)
+}
+func (m *MockStreamConnection) Close() error {
+	args := m.Called()
+	return args.Error(0)
+}
+func (m *MockStreamConnection) CloseChan() <-chan bool {
+	args := m.Called()
+	return args.Get(0).(<-chan bool)
+}
+func (m *MockStreamConnection) SetIdleTimeout(timeout time.Duration) {
+	m.Called(timeout)
+}
+
+// MockStream can be returned from a MockStreamConnection
 type MockStream struct {
 	mock.Mock
 	io.ReadWriteCloser
@@ -80,29 +91,6 @@ func (m MockStream) Identifier() uint32 {
 	return args.Get(0).(uint32)
 }
 
-// FIXME: docstring
-type MockStreamConnection struct {
-	mock.Mock
-	httpstream.Connection
-	headers http.Header
-}
-
-func (m *MockStreamConnection) CreateStream(headers http.Header) (httpstream.Stream, error) {
-	args := m.Called(headers)
-	return args.Get(0).(httpstream.Stream), args.Error(1)
-}
-func (m *MockStreamConnection) Close() error {
-	args := m.Called()
-	return args.Error(0)
-}
-func (m *MockStreamConnection) CloseChan() <-chan bool {
-	args := m.Called()
-	return args.Get(0).(<-chan bool)
-}
-func (m *MockStreamConnection) SetIdleTimeout(timeout time.Duration) {
-	m.Called(timeout)
-}
-
 // create an http request with the specified details
 func MockHttpRequest(method string, path string, headers map[string][]string, content string) http.Request {
 	return http.Request{
@@ -112,13 +100,4 @@ func MockHttpRequest(method string, path string, headers map[string][]string, co
 		ContentLength: int64(len(content)),
 		Body:          ioutil.NopCloser(bytes.NewBufferString(content)),
 	}
-}
-
-// assert that the content of the stream message coming from outputChan is equal to testSTring
-// this is a common pattern when testing the agent
-func AssertNextMessageHasContent(assert *assert.Assertions, outputChan chan smsg.StreamMessage, testString string) {
-	message := <-outputChan
-	content, err := base64.StdEncoding.DecodeString(message.Content)
-	assert.Nil(err)
-	assert.Equal([]byte(testString), content)
 }
